+++ packages/jest-runner/src/runTest.ts
@@ -144,6 +144,14 @@ async function runTestInternal(
     docblockPragmas,
     testPath: path,
   });
+
+  if (typeof environment.getVmContext !== 'function') {
+    console.error(
+      `Test environment found at "${testEnvironment}" does not export a "getVmContext" method, which is mandatory from Jest 27. This method is a replacement for "runScript".`,
+    );
+    process.exit(1);
+  }
+
   const leakDetector = config.detectLeaks
     ? new LeakDetector(environment)
     : null;